assignment-04/requirements.txt
fastmcp>=0.4.0
mcp[cli]>=1.0.0
prefab-ui>=0.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.32.0
feedparser>=6.0.11
defusedxml>=0.7.1
